Functions

GliderAnalysis.transect_xy — Function

transect_xy(longitude, latitude; inverse=false, case=:halifax_line, debug=0)

Convert longitude and latitude vectors to x and y vectors (in km), or do the reverse (if inverse=false). The transformation is governed by the case keyword (which has a default value that is useful along the Halifax Line on the Scotian Shelf, off Nova Scotia).

Arguments

longitude vector of longitudes

latitude vector of longitudes

Keywords

case either the symbol :halifax_line (which is the default) or a NamedTuple containing the elements longitude0 (the zero of new coordinate system), latitude0 (the zero of new coordinate system), angle (the counterclockwise angle of new coordinate system to an east-north system and srs (the coordinate transformation code). In the default case, the tuple defaults to (longitude0=-63.507773, latitude0=44.623249, angle=-59.1879, srs="EPSG:32620"). See Ref 1 for information on this default srs value.

References

https://epsg.io/32620 states the scope of EPSG:32620 to be navigation at medium accuracy, with application to latitudes from 0N to 84N and longitudes from 66W to 60W.

Examples

using GliderAnalysis, Plots, DataFrames, CSV

file = joinpath(dirname(dirname(pathof(GliderAnalysis))), "data", "sbloom_2023_traj.csv.gz")
traj = CSV.read(file, DataFrame);

xy = transect_xy(traj.longitude, traj.latitude);

fs = 7
KW = (framestyle=:box, tickdirection=:out, label=false, ms=1, guidefontsize=fs, tickfontsize=fs, titlefontsize=fs)
a = Plots.scatter(traj.longitude, traj.latitude, aspect_ratio=1.4, xlab="Longitude", ylab="Latitude", title=file; KW...)
b = Plots.scatter(xy[:, 1], xy[:, 2], aspect_ratio=1.0, xlab="Easting [km]", ylab="Northing [km]", title="Using transect_xy()"; KW...)
Plots.vline!([0], lwd=3, color=:red, label=false)
Plots.plot(a, b, layout=(1, 2), size=(800, 500))